requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pymysql==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.6
